+++ src/provider/decorationProvider.ts
@@ -1,4 +1,4 @@
-import { HttpFile, HttpSymbolKind, httpFileStore } from 'httpyac';
+import { HttpFile, HttpSymbolKind, HttpFileStore } from 'httpyac';
 import * as vscode from 'vscode';
 import { getConfigSetting, httpDocumentSelector } from '../config';
 
@@ -7,32 +7,29 @@ export class DecorationProvider {
   private subscriptions: Array<vscode.Disposable>;
   private decoration: vscode.TextEditorDecorationType;
 
-  constructor(context: vscode.ExtensionContext, httpFileEmitter: vscode.EventEmitter<{ httpFile: HttpFile, document: vscode.TextDocument }>) {
+  constructor(
+    context: vscode.ExtensionContext,
+    refreshCodeLens: vscode.EventEmitter<void>,
+    private readonly httpFileStore: HttpFileStore
+  ) {
     this.decoration = vscode.window.createTextEditorDecorationType({
       gutterIconPath: context.asAbsolutePath('./assets/gutter.svg'),
       gutterIconSize: '70%',
     });
 
     this.subscriptions = [
-      httpFileEmitter.event(({ document, httpFile }) => this.showGutterIcon(document, httpFile)),
-      vscode.window.onDidChangeActiveTextEditor(editor => {
-        if (editor && vscode.languages.match(httpDocumentSelector, editor.document)) {
-          const httpFile = httpFileStore.get(editor.document.fileName);
-          if (httpFile) {
-            this.setDecoration(httpFile, editor);
-          }
-        }
-      }, this)
+      refreshCodeLens.event(() => this.setEditorDecoration(vscode.window.activeTextEditor)),
+      vscode.window.onDidChangeActiveTextEditor(this.setEditorDecoration, this)
     ];
   }
 
-
-  private showGutterIcon(document: vscode.TextDocument, httpFile: HttpFile) {
-    const editor = vscode.window.visibleTextEditors.find(obj => obj.document === document);
-    if (editor) {
-      this.setDecoration(httpFile, editor);
+  private setEditorDecoration(editor: vscode.TextEditor | undefined) {
+    if (editor && vscode.languages.match(httpDocumentSelector, editor.document)) {
+      const httpFile = this.httpFileStore.get(editor.document.fileName);
+      if (httpFile) {
+        this.setDecoration(httpFile, editor);
+      }
     }
-
   }
 
   private setDecoration(httpFile: HttpFile, editor: vscode.TextEditor) {