+++ src/utils/httpRegionUtils.ts
@@ -1,9 +1,12 @@
 import * as vscode from 'vscode';
-import { httpFileStore, HttpRegionSendContext } from 'httpyac';
-import { initHttpClient } from '../config';
+import { HttpRegionSendContext, HttpFileStore } from 'httpyac';
 
 
-export async function getHttpRegionFromLine(doc: vscode.TextDocument | undefined, line: number | undefined) : Promise<HttpRegionSendContext | undefined> {
+export async function getHttpRegionFromLine(
+  doc: vscode.TextDocument | undefined,
+  line: number | undefined,
+  httpFileStore: HttpFileStore
+): Promise<HttpRegionSendContext | undefined> {
   const document = doc?.getText ? doc : vscode.window.activeTextEditor?.document;
   if (document) {
     const httpFile = await httpFileStore.getOrCreate(document.fileName, () => Promise.resolve(document.getText()), document.version);
@@ -12,7 +15,7 @@ export async function getHttpRegionFromLine(doc: vscode.TextDocument | undefined
       if (currentLine !== undefined) {
         const httpRegion = httpFile.httpRegions.find(obj => obj.symbol.startLine <= currentLine && currentLine <= obj.symbol.endLine);
         if (httpRegion) {
-          return { httpRegion, httpFile, httpClient: initHttpClient() };
+          return { httpRegion, httpFile };
         }
       }
     }